+++ src/Events/ModelWasBanned.php
@@ -9,32 +9,32 @@
  * file that was distributed with this source code.
  */
 
-namespace Cog\Ban\Events;
+namespace Cog\Laravel\Ban\Events;
 
-use Cog\Ban\Contracts\Ban as BanContract;
-use Cog\Ban\Contracts\Bannable as BannableContract;
+use Cog\Laravel\Ban\Contracts\Ban as BanContract;
+use Cog\Laravel\Ban\Contracts\Bannable as BannableContract;
 use Illuminate\Contracts\Queue\ShouldQueue;
 
 /**
  * Class ModelWasBanned.
  *
- * @package Cog\Ban\Events
+ * @package Cog\Laravel\Ban\Events
  */
 class ModelWasBanned implements ShouldQueue
 {
     /**
-     * @var \Cog\Ban\Contracts\Bannable
+     * @var \Cog\Laravel\Ban\Contracts\Bannable
      */
     public $model;
 
     /**
-     * @var \Cog\Ban\Contracts\Ban
+     * @var \Cog\Laravel\Ban\Contracts\Ban
      */
     public $ban;
 
     /**
-     * @param \Cog\Ban\Contracts\Bannable $bannable
-     * @param \Cog\Ban\Contracts\Ban $ban
+     * @param \Cog\Laravel\Ban\Contracts\Bannable $bannable
+     * @param \Cog\Laravel\Ban\Contracts\Ban $ban
      */
     public function __construct(BannableContract $bannable, BanContract $ban)
     {