+++ src/Providers/BanServiceProvider.php
@@ -9,20 +9,20 @@
  * file that was distributed with this source code.
  */
 
-namespace Cog\Ban\Providers;
+namespace Cog\Laravel\Ban\Providers;
 
-use Cog\Ban\Console\Commands\DeleteExpiredBans;
-use Cog\Ban\Contracts\Ban as BanContract;
-use Cog\Ban\Contracts\BanService as BanServiceContract;
-use Cog\Ban\Models\Ban;
-use Cog\Ban\Observers\BanObserver;
-use Cog\Ban\Services\BanService;
+use Cog\Laravel\Ban\Console\Commands\DeleteExpiredBans;
+use Cog\Laravel\Ban\Contracts\Ban as BanContract;
+use Cog\Laravel\Ban\Contracts\BanService as BanServiceContract;
+use Cog\Laravel\Ban\Models\Ban;
+use Cog\Laravel\Ban\Observers\BanObserver;
+use Cog\Laravel\Ban\Services\BanService;
 use Illuminate\Support\ServiceProvider;
 
 /**
  * Class BanServiceProvider.
  *
- * @package Cog\Ban\Providers
+ * @package Cog\Laravel\Ban\Providers
  */
 class BanServiceProvider extends ServiceProvider
 {