+++ tests/database/factories/BanFactory.php
@@ -9,8 +9,8 @@
  * file that was distributed with this source code.
  */
 
-use Cog\Ban\Models\Ban;
-use Cog\Ban\Tests\Stubs\Models\User;
+use Cog\Laravel\Ban\Models\Ban;
+use Cog\Tests\Laravel\Ban\Stubs\Models\User;
 use Faker\Generator;
 
 /* @var \Illuminate\Database\Eloquent\Factory $factory */