+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to `laravel-ban` will be documented in this file.
 
 ### Changed
 
+- All classes namespaces changed from `Cog\Ban\*` to `Cog\Laravel\Ban\*`  
 - Renamed database table `ban` to `bans`
 - Renamed database column `owned_by_id` to `bannable_id`
 - Renamed database column `owned_by_type` to `bannable_type`